fix(pruner): findPruneable short-circuits if estimated cutoff height is too close to last pruned
